backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.11
langchain==0.3.13
langchain-google-genai==2.0.7
langchain-community==0.3.13
python-dotenv==1.0.1
tavily-python==0.5.0
requests==2.32.3
pydantic==2.10.3
